requirements.txt
twscrape
pandas
numpy
sqlalchemy
sentence-transformers
faiss-cpu
hdbscan
scikit-learn
requests
python-dotenv

